requirements.txt
cryptography
pyjwt[crypto]
bcrypt
requests
